% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix_exp_dt}
\alias{matrix_exp_dt}
\title{Matrix exponential via eigen decomposition (small matrices)}
\usage{
matrix_exp_dt(A, dt)
}
\arguments{
\item{A}{detail}

\item{dt}{detail}
}
\description{
Matrix exponential via eigen decomposition (small matrices)
}
