% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsc_analyze.R
\name{plot.dsc_result}
\alias{plot.dsc_result}
\title{Plot DSC results}
\usage{
\method{plot}{dsc_result}(x, show_baseline = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{dsc_result}}

\item{show_baseline}{Logical; whether to show the baseline}

\item{...}{Additional graphical parameters}
}
\description{
Plot DSC results
}
