% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tga_analyze.R
\name{plot.tga_result}
\alias{plot.tga_result}
\title{Plot TGA results}
\usage{
\method{plot}{tga_result}(x, type = c("both", "tga", "dtg"), ...)
}
\arguments{
\item{x}{An object of class \code{tga_result}}

\item{type}{Plot type: "both", "tga", or "dtg"}

\item{...}{Additional graphical parameters}
}
\description{
Plot TGA results
}
