% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsc_analyze.R
\name{print.dsc_result}
\alias{print.dsc_result}
\title{Print DSC analysis summary}
\usage{
\method{print}{dsc_result}(x, ...)
}
\arguments{
\item{x}{An object of class \code{dsc_result}}

\item{...}{Ignored}
}
\description{
Print DSC analysis summary
}
