% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tga_analyze.R
\name{print.tga_result}
\alias{print.tga_result}
\title{Print TGA analysis summary}
\usage{
\method{print}{tga_result}(x, ...)
}
\arguments{
\item{x}{An object of class \code{tga_result}}

\item{...}{Ignored}
}
\description{
Print TGA analysis summary
}
